refactor: move lib/ to stores/ and utils/ folders, update all import paths
