+++ components/ui/section.tsx
@@ -1,5 +1,5 @@
 import { cva, type VariantProps } from "class-variance-authority";
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/theme";
 
 const sectionVariants = cva("py-10 lg:py-14", {
   variants: {
