+++ stores/app-state.ts
